+++ tests/events/EventFormComponents/ColorPicker.test.tsx
